+++ package.json
@@ -26,10 +26,11 @@
     "chart.js": "4.4.1",
     "core-js": "3.36.0",
     "dropzone": "6.0.0-beta.2",
-    "postcss": "^8.4.47",
     "pinia": "^2.2.4",
     "pinia-plugin-persistedstate": "^4.1.1",
+    "postcss": "^8.4.47",
     "quill": "^2.0.2",
+    "three": "^0.169.0",
     "v-calendar": "^3.1.2",
     "vue": "3.4.19",
     "vue-count-to": "1.0.13",
