+++ src/examples/FooterExample.vue
