+++ src/examples/FooterExample2.vue
