+++ src/router/index.js
@@ -1,12 +1,12 @@
 import { createRouter, createWebHistory } from "vue-router";
 import Dashboard from "../views/Dashboard.vue";
-import Tables from "../views/Tables.vue";
-import Billing from "../views/Billing.vue";
-import VirtualReality from "../views/VirtualReality.vue";
-import RTL from "../views/Rtl.vue";
-import Profile from "../views/Profile.vue";
-import Signup from "../views/Signup.vue";
-import Signin from "../views/Signin.vue";
+import Tables from "../views/Mainpage/Tables.vue";
+import Billing from "../etc/Billing.vue";
+import VirtualReality from "../etc/VirtualReality.vue";
+import RTL from "../etc/Rtl.vue";
+import Profile from "../views/Profile/Profile.vue";
+import Signup from "../views/Sign/Signup.vue";
+import Signin from "../views/Sign/Signin.vue";
 
 const routes = [
   {
