+++ src/views/PageCustom/Configurator.vue
